backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
pydantic-settings==2.2.1
slowapi==0.1.9
google-generativeai==0.5.4
groq==0.9.0
sendgrid==6.11.0
resend==1.1.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
ruff==0.4.4
